+++ resources/cli-templates/install/app/Filters/VeilData.php
@@ -22,7 +22,7 @@ class VeilData extends Filter implements FilterInterface
 
     public function isActive(): bool
     {
-        return $this->container->has('veil');
+        return $this->container->has('Bayfront\Veil\Veil');
     }
 
     /**